+++ src/classy_blocks/optimize/clamps/curve.py
@@ -98,4 +98,4 @@ class RadialClamp(ClampBase):
 
     @property
     def initial_guess(self):
-        return [0]
+        return [0.0]