+++ src/classy_blocks/optimize/optimizer.py
@@ -9,6 +9,7 @@ import numpy as np
 import scipy.optimize
 
 from classy_blocks.base.exceptions import OptimizationError
+from classy_blocks.construct.flat.sketch import Sketch
 from classy_blocks.construct.flat.sketches.mapped import MappedSketch
 from classy_blocks.construct.operations.operation import Operation
 from classy_blocks.mesh import Mesh
@@ -51,7 +52,7 @@ class OptimizerConfig:
     # as passed to scipy.optimize.minimize
     clamp_tol: float = 1e-3
     # additional options passed to Scipy's minimize method,
-    # depending on chosen algotirhm; see documentation of scipy.optimize.minimize
+    # depending on chosen algorithm; see documentation of scipy.optimize.minimize
     # and specifically the chosen algorithm
     options: dict = field(default_factory=dict)
 
@@ -109,6 +110,10 @@ class OptimizerBase(abc.ABC):
         initial_params = copy.copy(clamp.params)
 
         def fquality(params):
+            if clamp.bounds is not None:
+                for i, b in enumerate(clamp.bounds):
+                    params[i] = np.clip(params[i], b[0], b[1])
+
             clamp.update_params(params)
             return self.grid.update(junction.index, clamp.position)
 
@@ -172,11 +177,10 @@ class OptimizerBase(abc.ABC):
 
         k = -np.log(1 - (threshold - start_relax) / (threshold - start_relax + VSMALL))
 
-        # Normalize iteration to [0, 1]
+        # normalize iteration to [0, 1]
         t = iter_no / target_iter
 
         # increase the factor slowly at the beginning and quicker at the end
-        # Slow start, fast finish
         value = start_relax + (threshold - start_relax) * (1 - np.exp(-k * (t**3)))
 
         return value
@@ -209,6 +213,9 @@ class OptimizerBase(abc.ABC):
         for i in range(self.config.max_iterations):
             iter_record = self._optimize_iteration(i)
 
+            if iter_record.abs_improvement < 0:
+                # can happen during the relaxed iterations
+                continue
             if iter_record.abs_improvement < self.config.abs_tol:
                 orecord.termination = "abs"
                 break
@@ -261,15 +268,24 @@ class ShapeOptimizer(OptimizerBase):
 
 
 class SketchOptimizer(OptimizerBase):
-    def __init__(self, sketch: MappedSketch, report: bool = True):
+    def __init__(self, sketch: Sketch, report: bool = True, merge_tol: float = TOL):
         self.sketch = sketch
 
-        grid = QuadGrid(sketch.positions, sketch.indexes)
+        grid = QuadGrid.from_sketch(sketch, merge_tol=merge_tol)
 
         super().__init__(grid, report)
 
     def _backport(self):
-        self.sketch.update(self.grid.points)
+        if isinstance(self.sketch, MappedSketch):
+            self.sketch.update(self.grid.points)
+            return
+
+        # take faces and points from QuadGrid
+        for face_index, face in enumerate(self.sketch.faces):
+            point_indexes = self.grid.addressing[face_index]
+            for corner_index, point_index in enumerate(point_indexes):
+                point = self.grid.points[point_index]
+                face.points[corner_index].position = point
 
     def auto_optimize(
         self,